% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{as.data.frame.mandelbrot}
\alias{as.data.frame.mandelbrot}
\title{Convert Mandelbrot object to data.frame for plotting}
\usage{
\method{as.data.frame}{mandelbrot}(mandelbrot)
}
\description{
Converts objects produced by \code{\link[mandelbrot]{mandelbrot}}
to tidy data.frames for use with ggplot and other tidyverse packages.
}
