% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{mandelbrot}
\alias{mandelbrot}
\title{Calculate the Mandelbrot set}
\usage{
mandelbrot(xlim = c(-3, 1), ylim = c(-1.8, 1.8), resolution = 600,
  iterations = 50)
}
\arguments{
\item{xlim}{limits of x axis (real part)}

\item{ylim}{limits of y axis (imaginary part)}

\item{resolution}{either an integer \eqn{n} for \eqn{n^2} pixels
or a list with x and y components specifying the resolution
in each direction (e.g. \code{list(x = 500, y = 500)})}

\item{iterations}{maximum number of iterations to
evaluate each case}
}
\description{
Generates a view on the Mandelbrot set
using an underlying C function
}
\details{
Original code by Mario dos Reis, September 2003.
}
\references{
\url{http://people.cryst.bbk.ac.uk/~fdosr01/Rfractals/index.html}
}
\author{
Mario dos Reis
}
