% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{mandelbrot_palette}
\alias{mandelbrot_palette}
\title{Generate palette suitable for coloring a set}
\usage{
mandelbrot_palette(palette, in_set = "black")
}
\description{
Takes a simple palette and expands / oscillates
it for use with Mandelbrot sets.
}
\examples{
blues <- RColorBrewer::brewer.pal("Blues", 9)
cols <- mandelbrot_palette(blues, in_set = "grey70")

spectral <- RColorBrewer::brewer.pal("Spectral", 11)
cols <- mandelbrot_palette(spectral)

view <- mandelbrot(x = c(-0.8438146, -0.8226294),
  y = c(0.1963144, 0.2174996), iter = 500)
image(-1/frac$z, col = spectral)

}
