% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{as.data.frame.mandelbrot}
\alias{as.data.frame.mandelbrot}
\title{Convert Mandelbrot object to data.frame for plotting}
\usage{
\method{as.data.frame}{mandelbrot}(x, ...)
}
\arguments{
\item{x}{a Mandelbrot set object produced by \code{\link[mandelbrot]{mandelbrot}}}

\item{...}{ignored}
}
\value{
a 3-column \code{data.frame}
}
\description{
Converts objects produced by \code{\link[mandelbrot]{mandelbrot}}
to tidy data.frames for use with ggplot and other tidyverse packages.
}
\examples{

mb <- mandelbrot()
df <- as.data.frame(mb)
head(df)

}
