% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{mandelbrot_palette}
\alias{mandelbrot_palette}
\title{Generate palette suitable for coloring a set}
\usage{
mandelbrot_palette(palette, in_set = "black")
}
\description{
Takes a simple palette and expands / oscillates
it for use with Mandelbrot sets.
}
\examples{
view <- mandelbrot(xlim = c(-0.8438146, -0.8226294),
  ylim = c(0.1963144, 0.2174996), iter = 500)

blues <- RColorBrewer::brewer.pal(9, "Blues")
cols <- mandelbrot_palette(blues, in_set = "white")
image(log(view$z), col = cols, axes = FALSE)

spectral <- RColorBrewer::brewer.pal(11, "Spectral")
cols <- mandelbrot_palette(spectral)
image(-1/view$z, col = cols, axes = FALSE)

}
