% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{plot.mandelbrot}
\alias{plot.mandelbrot}
\title{Plot a Mandelbrot set using base graphics}
\usage{
\method{plot}{mandelbrot}(mandelbrot, col = mandelbrot_palette(c("white",
  grey.colors(50))), transform = c("none", "inverse", "log"), ...)
}
\arguments{
\item{mandelbrot}{}
}
\description{
Draws coloured set membership using \code{image}.
}
